lxd/device/usb: Include bus and dev num in device name
If there are multiple devices with the same product and vendor ID, the
device names have the suffixes "-0", "-1", etc. This works fine for
static devices, but can cause issues when supporting hotplugging due to
the ordering of these devices.

It's important that devices are removed from QEMU once the are
physically removed. That's because if they are plugged in again, the
internal QEMU device name must not exist. For this, they need to be
identified correctly. An incremental suffix value therefore won't
work.

This replaces the incremental value suffix with the bus and device
number which makes the device identifiable.

Signed-off-by: Thomas Hipp <thomas.hipp@canonical.com>