lxd/instance/qemu: Allow using external firmware or kernel
With this change, the -bios or -kernel options inside of raw.qemu will
cause LXD to log a warning message and then proceed to disable the
normal firmware setup.

This allows for booting LXD off of an external firmware file or directly
into a specific kernel. Such firmware or kernel will likely need a
matching raw.apparmor entry so that QEMU can actually access them.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>