lxd/storage/drivers/btrfs: Change how subvolumes are received
When subvolumes were received, they were made read-write as they would
be moved to their final destination right after being received. This
doesn't work when performing a refresh.

This removes the property change, making all received subvolumes
readonly. The caller is responsible for making the subvolume read-write
if they wish to move or delete it. It also determines the latest
received path by comparing the directory content before and after the
reception.

Signed-off-by: Thomas Hipp <thomas.hipp@canonical.com>