shared: allow EOPNOTSUPP from llistxattr()
Some filesystems don't support llistxattr() for various reasons. Let's
interpret this as a set of no xattrs, instead of an error.

The goal here is to be able to use the idmap shifting code on xattr-less
filesystems, which is not currently possible, since there is the following
call graph:

github.com/lxc/lxd/shared/idmap.(*IdmapSet).doUidshiftIntoContainer.func1
-> github.com/lxc/lxd/shared/idmap.GetCaps
-> github.com/lxc/lxd/shared.GetAllXattr
-> github.com/lxc/lxd/shared.llistxattr

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>