lxd/storage/drivers/zfs: Close stderr after copy
In an error case, if `ioutil.ReadAll(stderr)` ran before the goroutine
ended, `ReadAll` would hang forever. It wouldn't happen every time,
but if called often enough it would eventually hang. This would make
it impossible for the source to clean up their instances since the
resource is busy.

This closes stderr at the end of the goroutine, making it impossible for
`ReadAll` to hang.

Signed-off-by: Thomas Hipp <thomas.hipp@canonical.com>