lxd/instance/lxc: Better handle missing apparmor
When `lxc.apparmor.profile` isn't specified, liblxc if it detects kernel
support will set the lxc-container-default-with-cgns profile. But that
profile won't be present if the apparmor tools are missing.

So as LXD knows better about the state of apparmor, have it tell liblxc
to completely back off by seting the profile to unconfined when apparmor
profiles aren't expected to be loaded.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>