lxd: Remove old clustering tests
Those were disabled for the longest time due to a dqlite issue and are
all testing APIs we have long deprecated (the old cluster joining method).

They were also meant to primarily test handling of the MemberConfig and
old join flow which has changed a lot and isn't something that can so
easily be mocked.

Our system tests cover everything in here and are what we've actually
been relying on for years now.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>