lxd/instance/qemu: Switch to socket for console
This removes some QMP calls, aligns the handling with what we have for
SPICE and it eliminates a potential source of race conditions as it
doesn't rely on the startup-time PTS path still being correct when
accessed at a later time.

This change unfortunately means that text console sessions on existing
VMs will break and require a VM reboot to switch the config.

It shouldn't be a huge deal though as normal exec sessions as well as
VGA consoles will still operate properly.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>