lxd/storage/drivers/btrfs: Delete volume after receiving new one
The main subvolume was removed before receiving the new one. In case of
an error while receiving, the subvolume would not exist but the database
entry would. Deleting the instance would therefore fail, requiring the
user to manually clean up that mess.

This moves the subvolume deletion to after the new subvolume has been
received. In case of an error while receiving, the old subvolume would
still exist.

Signed-off-by: Thomas Hipp <thomas.hipp@canonical.com>