lxd/storage/drivers/btrfs: Set Received UUID
When making the received subvolume read-write before moving it to its
final location, the "Received UUID" information is lost (by design).
However, this causes issues when performing incremental streams and
can result in a "cannot find parent subvolume" error.

Setting the "Received UUID" field to the value of the received
subvolume (before making it rw) solves this issue.

Signed-off-by: Thomas Hipp <thomas.hipp@canonical.com>