doc: add a version switcher to the docs
Add a version switcher that displays the versions defined in
the versions.json file (from master) in a dropdown.
Selecting a different version will switch to the corresponding
file in the other documentation version, or to the start page
of that documentation version if the file doesn't exist in the
other version.

The script should use the versions.json file from the master
branch for all documentation versions, because this file
always has the latest information on which versions are relevant.
This way, if we add or remove a version, we don't need to update
old documentation sets to do the same change, but it will happen
automatically.

Signed-off-by: Ruth Fuchss <ruth.fuchss@canonical.com>