lxd/backup: Updates backupWriteIndex to use pool.GenerateInstanceBackupConfig
This removes duplication and has the side effect of including the contents of
the instance's on disk backup.yaml in index.yaml.

This is desirable as it will mean that in the future we can potentially create
the storage volume DB records and on-disk storage volumes at the correct size
and config before unpacking a backup into it.

This is because for optimized backups, the instance's backup.yaml file is not
accessible in the tarball as its part of the binary blob from the storage driver.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>