lxd/migration/migration/volumes: Adds InfoResponse struct
This is used as a response back to the Info index header frame.
Right now it doesn't contain anything useful, but the intention here is to extend it with
additional info from the target about which frames (such as snapshots) it wants to receive.
The difference compared to the protobuf based negotiation that takes place at the moment
is that this response frame will be sent back after the index frame has been negotiated
and sent, which allows the target to use the full contents of the index header to decide
which snapshots it wants.

My intention is that this will be useful in solving the issue of figuring out which
snapshots need syncing if they have the same name.

See https://github.com/lxc/lxd/issues/10436

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>