lxd/db/cluster/instances.mapper: Generate cluster package methods for instances
This is quite incomplete, because fully migrating Instances over to the
database is a bit too complex at this time. Unfortunately this means
Instance, Profile, and InstanceProfile types will simultaneously exist
for a little while both under the db package and the cluster package,
until instances can be fully migrated over to the cluster package.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>