client: Move image profile override to ImageCopyArgs
The current implementation has the issue of attempting to pass in a list
of profiles to server which may not have the feature to begin with.

Instead use ImageCopyArgs to pass in the list of profiles and if that is
set, then perform the extension check. If it's not set, then clear any
profile that may be listed on the image prior to copy.

This not only fixes the situation of interacting with a server which
doesn't support overriding image profiles but also avoids attempting to
use potentially non-existent profiles during a normal copy.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>