lxd/storage/drivers/zfs: Fix VM migration
The createVolumeFromMigrationOptimized function calls
CreateVolumeFromMigration for VMs. This is wrong, as the receiver
then expects the migration headers to be sent twice which the sender
doesn't do. It results in the receiver failing to decode the migration
header (JSON) as the received data at that point is actually from
`zfs send`.

This changes createVolumeFromMigrationOptimized to call itself for VM
block filesystems.

Fixes #104435.

Signed-off-by: Thomas Hipp <thomas.hipp@canonical.com>