lxd/storage/drivers/btrfs: Don't rely on passed snapshots
The passed snapshots don't always include all which need to be
transferred. This can happen with optimized copies (btrfs and zfs) as
the backend has no knowledge of the UUIDs/GUIDs of datasets or
subvolumes, and can therefore pass an incorrect/incomplete list
of snapshots.

This lets the drivers negotiate which snapshots actually need to be
transferred.

Signed-off-by: Thomas Hipp <thomas.hipp@canonical.com>