lxd-user: Move away from atomic
This implements the same change as we did to lxd/main_forkfile.go recently.
We've had very weird issues with Go's atomic implementation and handling
of accept on sockets.

This hasn't yet been an issue with lxd-user, but given both forkfile and
lxd-user received an identical implementation of atomics, it makes sense
to keep them lined up.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>