lxd-user: Use getent instead of os/user
The Go os/user effectively parses /etc/passwd which is incompatible with
external authentication through nss. It's possible to link it against
libnss to get access to those, but it's still problematic for static
builds or in environments like the snap.

So instead, just call `getent passwd` and parse the output.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>