lxd/db/instances: Use InstanceToArgs in InstanceList
This reworks InstanceList a bit to both reduce the number of
queries and use InstanceArgs to get all the data for an instance.
As InstanceArgs now holds all the api.Profiles each instance's profiles,
InstanceList has been simplified to just fetch the database instances,
convert them to InstanceArgs, and convert their projects to api.Project.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>