lxd/instance/drivers/driver/common: Allow instances to be created and updated with conflicting devices
This allows instances with conflicting devices to be:

 - Created from as a copy
 - Restored from backup
 - Restored from snapshot

Which would otherwise fail, and would be difficult to correct the problem without manually modifying the database or backup file.
This then allows the user to manually the fix the conflict before the instance can be started.

We do not run the deviceAdd function when a device conflicts with an existing device because that could cause
the existing device's setup on the system to be disrupted (think conflicting static DHCP leases being created).

Fixes #10114

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>