lxd/db/generate/db/stmt: Allow overriding table name with config key
Sometimes the generator just isn't smart enough to parse table names
from go struct names, so this adds a config key to override the table
name with a specified one.

An example of this is "cluster_members". The generator parses this like
it would for "instances_profiles", resulting in "clusters_members".
Blocking "cluster" from being pluralized would work, but it's a
completely appropriate pluralization that might cause issues if we ever
have a table actually representing more than one cluster. Having the
config key bypasses this whole dilemma.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>