lxd/db/generate/db/method/v2: Limit generated method references to same package
There's no way to know without parsing the whole package whether a
function is defined locally in the package that is calling lxd-generate,
or under the -d <database> specified package name. This results in
inconsistent behaviour, so this commit removes appending of the -d
<database> package name to helper methods. If we want to reference those
helpers we can just generate new helpers in the current package.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>