doc: Update cloud-init.md with user-data examples
Context
==
I found the current documentation to be very lacking in terms of a clear explanation of how to apply the `user-data` configuration options, and how to confirm that the `cloud-init` process had received the data. It was only through a lot of searching that I found out that `#cloud-config` is actually not just a comment, but actually required for `cloud-init` to even process that set of piped configuration keys.

Approach
==
This change
* Adds a small "how to" for setting up a new profile for testing
* Adds documentation on the pipe syntax that's needed to pipe the configuration to cloud-init
* Adds documentation for the specific syntax required to get `cloud-init.user-data` to work (`#cloud-config`)
* Adds documentation on where to check on the instance to see if the user-data was passed in properly
* Adds examples of using the `user-data` section to take actions in an instance

Errata
==
I acknowledge that v5 is meant to bring a documentation rewrite per the comment on the YouTube video linked from this document, but I don't think that's a good reason to not have some clear examples in the documentation right now. Google searches tend (for me) to find documentation from 2018 that use the old config syntax (user.user-data), and that documentation doesn't say `#cloud-config` is crucial.

Testing of instructions
==

With the `#cloud-config` line in the `cloud-init.user-data` block:
```
$ ls -l /var/lib/cloud/instance | grep -E 'cloud-config.txt|user-data.txt'
-rw------- 1 root root  149 Jul 16 08:47 cloud-config.txt
-rw------- 1 root root  117 Jul 16 08:47 user-data.txt
```
`/var/log/syslog` confirms `apt upgrade` and `apt install` were run.

Without the `#cloud-config` line:
```
$ ls -l /var/lib/cloud/instance | grep -E 'cloud-config.txt|user-data.txt'
-rw------- 1 root root    0 Jul 16 08:09 cloud-config.txt
-rw------- 1 root root  103 Jul 16 08:09 user-data.txt
```

`/var/log/syslog` confirms that none of the upgrades etcetera ran.

Signed-off-by: Duncan <github.com@cricalix.net>