lxd/db/generate/db/method: Support multiple filter structs
Adds support for passing multiple <entity>Filter structs to
Get<entities> methods.

This implementation creates a new statement from the SQL string of a
previously existing prepared statement, and modifies its WHERE clause to
a WHERE IN clause, with the number of parameters matching the number of
its corresponding <entity>Filter struct field.

I had previously tested preparing statements with a preset large number
of WHERE-IN parameters, and using nil values for the unused parameters,
but this had a very negligible effect on query performance, at the cost
of much harder to read code. That implementation also struggled a lot
with https://github.com/canonical/dqlite/issues/377

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>