lxd/db/storage/volumes: Updates GetStoragePoolVolumes to return []StorageVolume
Adds new StorageVolume type that embeds api.StorageVolume but has additional fields
for ID and Project. This is similar to the DB generator types, but without the
additonal complexity needed for converting from the DB type to the API type, as
it is embedded directly.

Returning a slice of this new type is preferable to complicated maps as it
allows us to predefine the resulting []api.Storage slice before its returned.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>