lxd/storage/volumes/utils: Updates storagePoolVolumeUsedByGet to accept a db.StorageVolume
Also clarifies that the project passed into it should be the requested project name on the URL,
rather than the effective project name of the storage volume.

Passes the volume's project into `VolumeUsedByInstanceDevices` so that it correctly matches
against instances in a non-default project (without custom volume feature) using a volume
from the default volume, when the requested project isn't default.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>