lxd/network/driver_ovn: Fixes impossibility to set MTU for a pair of veths connecting uplink and ovs-system bridges
If an existing Linux native bridge is used as an uplink for OVN,
the (managed) uplink interface type is "physical". So it hasn't
"bridge.mtu" option, only "mtu" one. But only "bridge.mtu" option
of uplink interface is honored by the "startUplinkPortBridgeNative"
function. Because of this, the interfaces of the veths pair
connecting the uplink and the ovs-system bridges always gets MTU
value 1500. When such interface is added to the uplink bridge,
the MTU of the uplink bridge is automatically reduced to 1500
even if the physical interface has a higher MTU.

Signed-off-by: Andrei Azhoichyk <azh.andrei@gmail.com>