idmap: handle llistxattr returning EOPNOTSUPP
When shifting a file into idmapset range, we list the xattrs
so we can restore the filecaps if needed.  However, the llistxattr
system call we use for that can return EOPNOTSUPP.  In that case
we should assume an empty xattr set, rather than return an error
and failing the whole file (and container) uid-shift operation.

Signed-off-by: Serge Hallyn <serge@hallyn.com>