GetAllXattr: handle overlay filtering
When calling listxattr with sz 0, overlay does not filter the
results.  When subsequently calling it with non-zero sz, it
will filter out some xattrs.  If all xattrs were filtered out,
then the next result will be 0.  In this case, we were continuing
on to call getxattr on xattr=="", and getting back an ERANGE,
after which we returned ERANGE for the whole ShiftFile operation.

So if listxattr returns 0 on our second call, simply return with
no error.

Signed-off-by: Serge Hallyn <serge@hallyn.com>