lxd/network/driver/bridge: Consider a dnsmasq start failure as a network start failure
This will allow the network start logic to kick in, retrying to start the network periodically.
It will also mean that a warning is logged in `lxc warning list` as well as the LXD log.

The firewall is cleared at the begining of the `Start()` function, but is not applied until
after dnsmasq has been started. But there are lots of other reasons this `Start()` function
could return prematurely so is not going to leave the firewall in any more of an inconsistent
state than is already possible. Additionally we only apply the firewall rules in their
entirety later on so its not like only some of the firewall rules will be applied.

Fixes #10874

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>