test: Adds LXD_REQUIRED_TESTS support
This allows for tests to return after setting TEST_UNMET_REQUIREMENT non-empty.
This will then log the test as having been skipped unless the test name is in
the LXD_REQUIRED_TESTS list in which case it will be considered a failure.

In either case the value of TEST_UNMET_REQUIREMENT will be printed to aid
fixing the problem.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>