lxd/db/generate/db/method: More efficient AssociationTable queries
The `GetMany` function generated for an AssociationTable type (so
InstanceProfiles and CertificateProjects) first fetched a list of
IDs, then initiated one query for each ID to retrieve the entities.
This was pretty inefficient.

With lxd-generate now supporting variadic filters, this
function has been updated to accept a list of EntityFilter for the
returned entity. This allows us two things:

1) To filter on the columns of the entity we want (except IDs, which
   will be supplied internally by the function)

2) Reduce the total number of queries to 2 (One for all the IDs, one
   for all the entities) by creating a set of filters for each
   returned ID, with all the caller supplied filters as well as the
   entity ID.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>