lxd/instance/drivers/load: Updates validDevices to accept local and expanded config
So it can validate them using differently.

Specifically, if the device is not supported by the instance type then this is considered
a validation failure for local devices, but not for expanded devices from profiles.

This is to allow mixed instance type profiles where some devices are only supported by
specific instance types.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>