lxd/instance/drivers/driver/qemu: Do not attempt to mount & unmount in generateAgentCert
We expect the callers of generateAgentCert to have ascertained whether the VM is running
(and therefore mounted) before calling this function.

By making this change it prevents a mount/unmount race when LXD starts up during the
`devicesRegister` function call.

During this function call, each instance has its `IsRunning()` function called to check
if it needs its devices registered. For `disk` devices of running instances its important
that their `Register`functions are called on LXD start in order to initialise the mount
counter for that disk (so that a subsequent unmount call isn't incorrectly run if LXD
incorrectly thinks that the disk volume isn't in use anymore).

For VMs the initial call to `IsRunning()` is also important as that triggers a reconnection
to the QEMU QMP monitor in the running VM and allows for gathering the lxd-agent run status.

The mount/unmount race before the `disk` devices have been registered is caused because the
`IsRunning` call initialises the `getMonitorEventHandler` and causes it to call the
`advertiseVsockAddress` function, which in turn uses `generateAgentCert` as part of creating
a connection to the lxd-agent over vsock.

The `getMonitorEventHandler` handler is run in a separate goroutine and thus it can race
the `disk` device's `Register` function (which would initialise the mount counter so that
an unmount call would not be issued as the volume would be considered in use).

If the `generateAgentCert` function finishes before the `disk` device's `Register` function
then an unmount call is incorrectly attempted on the running VM's root disk causing delays
to LXDs start up (as the unmount call blocks for several seconds).

By removing the mount management from the `generateAgentCert` function it ensures that no
early unmount attempt can occur, and doesn't affect the operation of the function as it
is required that the VM be running (and thus mounted) in order to connect to the lxd-agent
anyway.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>