lxd/storage/backend/lxd: Update DeleteImage to not use GetImageFromAnyProject
We can ascertain the content type of the volume from the storage volume record,
rather than relying on the image DB record existing.

This makes this function more resilient as it removes the dependency of having
an associated image DB record, which in some cases where the DB is not in a
consistent state, can prevent the storage pool from being deleted.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>