lxd/instance/drivers: Fixes container initiated instance-shutdown lifecycle event
Because there are 2 container stop hooks (onstopns and onstop) the instance initiated operation lock
was being created on onstopns, but the instance-shutdown lifecycle event was only being generated in
the onstop hook. Because of this the operation lock already existed (because it was created by the
onstopns hook) by the time onstop hook run and the instance-shutdown lifecycle event never fired.

This has been fixed by adding an instance initiated marker on the operation lock itself which can be
propagated between stop hooks, allowing for the final onstop hook to detect an instance initiated
stop and correctly trigger the instance-shutdown lifecycle event.

VMs were never affected because they don't have the same stop hooks that containers have.

Fixes https://discuss.linuxcontainers.org/t/container-shutdown-event-not-raised-after-executing-halt-in-container/7191

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>