lxc/init: Improve cmdInit device override to work nicely with --storage and --network flags
If these flags are used then a local device is generated, but this was being replaced with
a device from the profiles if combined with the --device flag.

Now we check if the device being overriden is already defined locally and if so then we
override that device instead.

As an optimisation we now also only load the profile(s) if a device being overriden isn't
present in the local request. This way if we are using `--storage` or `--network` and only
overriding those local devices then there's no need to load the profile(s) from the server.

Also improves some errors.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>