lxd/storage/drivers/driver/lvm: Overrides roundVolumeBlockSizeBytes function
This function uses the LVM volume group's extent size for rounding.
This allows for instances published to images from an LVM volume to be used
as the basis for new images using the original disk size setting.

Although this does not guarantee that an image published from an LVM volume
will be usable for instances on non-LVM pools using the same disk size
setting, as they may use different disk size rounding methods.

In those cases the new instances will need to use a larger disk size setting.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>