lxd/network/network/utils: Adds network type to NIC connected to network logic
Allows for correctly detecting that `ovn` NICs can only use `network` and not `nictype`.
This prevents an issue where a `macvlan` NIC could be incorrectly considered connected to
an OVN network if it was using `parent` property connected to an unmanaged bridge of the
same name as an OVN network in a non-default project.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>