lxd/main/forkdns: Updates serversFileMonitor to only watch for inotify.InMovedTo event
This fixes a regression caused when I moved this to use the inotify package as it was
watching all events, including the open and read events causing an event loop.

This also fixes a small inefficiency that caused the forkdns server to reload its
servers file twice because it was triggering both on writes to the servers file
and the temporary file that is generated first.

We now only watch for files being renamed to the main servers list file.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>