lxd/instance/drivers/driver/qemu: Fix multi-queue NIC support in addNetDevConfig
This was regressed as part of https://github.com/lxc/lxd/pull/11075

Now we can have hot-pluggable CPUs we need to query the state of the VM at NIC
add time to get a list of CPUs in the guest.

This allows us to dispense with the cpuCount argument to addNetDevConfig, and
by extension to simplify return value of addCPUMemoryConfig as well.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>