lxd/instance/drivers/driver/qemu: Set reboot action to shutdown
So that LXD only needs to handle the SHUTDOWN event from the guest.
It can then use the reason code inside that event to detect a guest initiated reboot
and restart the guest cleanly.

This is cleaner than trying to catch the RESET event and then powering off the VM to restart.
It also means that restart logic is only handled in the onStopHook() rather than in two places.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>