lxd/instance/drivers: Fixes delete of ephemeral VM on stop
Moves the IsRunning check out of the internal `delete()` function and into the
exported `Delete()` function. This avoids failing the the `delete()` call from
the `onStop()` function when trying to delete an ephemeral VM on stop.

Although containers were not affected (because their notion of running doesn't
include an ongoing Stop operation that VMs do) I've also moved the check in
the LXC driver to the same place for consistency.

Fixes #11261

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>