lxd/instance/drivers/driver/lxc: Extend operation lock in onStopNS and onStop
Ensure the operation lock is extended as quickly as possible as soon as it becomes clear
that the container has stop (as we are then out of the client provided shutdown timeout condition).

Also, don't wait for `d.IsRunning()` as container will already be considered stopped by then, and
this avoids another call into liblxc to get the container's state.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>