lxd/device/nic/macvlan: Enable all multicast processing on VM macvtap host interface
Required for correct operation of IPv6 NDP.
Previously this was only working for SLAAC addresses because the host interface still
had IPv6 enabled and so was generating its own link-local address which was in the
same multicast group required for SLAAC address operation.

However if the guest used a non-SLAAC address (static or DHCPv6 stateful) because the
host interface did not know about this the NDP solication didn't work and the address
was unreachable.

This disables IPv6 on the host-side to ensure there are no IPv6 addresses at all on the
host interface, and instead enables all multicast processing to allow the guest to
configure any IP it wants.

This aligns with the macvlan behaviour for containers, where because we pass the host
interface directly into the container, any addresses added to the guest interface are
known about by the host and the correct multicast groups are joined for NDP to work.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>