lxd: hotplug mounts
For centuries we've worked ourselves to the bone with mount propagation
to hotplug mounts. No more! The new mount api does have proper support
for moving a mount into a container. I've had that on my ToDo for a
while but due to recent events this has been bumped up on my priority
list.

So here it is. A way to simply have LXD prepare a mount with exactly the
option we want and then to hand that fd to the container and the
container can just attach it to the correct place. No more pointless
mount propagation shenaningans.

I would also like to call out that this has some elegant properties. For
example, every filesystem that supports idmapped mounts can be exposed
to a container completely idmapped. The filesystem can be directly
created with the idmap applied. Neitehr does the filesystem have to
appear anywhere before applying the idmapping nor are any additional
mounts required before applying the idmapping.

Signed-off-by: Christian Brauner (Microsoft) <brauner@kernel.org>