lxd/storage/drivers/zfs: don't force atime to on
zfs defaults to using atime=on and relatime=off and man 7 zfsprops:

> relatime=on|off
>   Controls the manner in which the access time is updated when atime=on is set.

To align with other FSes, we want relatime=on. By not forcing atime to on, we
get the relatime behavior and we also allow an admin to disable the feature by
setting atime=off on the dataset itself or on a parent.

This also fixes commit d3e3b3df775f989a6f

Signed-off-by: Simon Deziel <simon.deziel@canonical.com>