lxd/storage/drivers/zfs/volumes: Freeze source filesystem volume when taking temporary snapshot in CreateVolumeFromCopy
When not in "allow inconsistent" mode, we must sync and freeze the source filesystem (if mounted) before taking the temporary
snapshot, otherwise there may unwritten pages that haven't made it to the underlying zvol and/or there could be ongoing
changes being made to the filesystem.

We try and keep the frozen time short by immediately unfreezing after the temporary snapshot has been taken.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>