lxd/instances: Don't use notify for targetting
When creating an instance using an alternative target whether specified
by the user or as the result of internal placement logic, we were
forwarding the request to that target with the notify bit on.

This would then cause the target server to think this was a legitimate
notification, that is, the completion of a cluster internal instance
relocation.

The result was some really odd errors, either complaining that the
instance didn't exist (when on Ceph) or that the source instance name
must be provided for a cluster instance move.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>